Imported Previous Version
Imported the last version of the project from before using GitHub.

Currently, the project requires quite unsafe definitions of x and potentialEnergies to run effectively, I plan to fix this by making a class called System with properties of particleEnergy, particleMass, x, potentialEnergies.

This class will have all of the methods of underlyingFunctions.jl and hence x and potentialEnergies will be defined in a safer way.